refactor: change structure of app layer
