+++ src/app/commands/mod.rs
@@ -1,0 +1,2 @@
+pub mod unauth;
+pub mod user_login;
